libprotoident

+++ src/main/scala/ru/laboshinl/tractor/TractorPayload.scala
@@ -3,4 +3,4 @@ package ru.laboshinl.tractor
 /**
  * Created by laboshinl on 9/16/16.
  */
-class TractorPayload(var startPos : Int, var stopPos: Int) extends Serializable
+class TractorPayload(var startPos : Long, var stopPos: Long) extends Serializable
